+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.4] - TESTING
+
+### Added
+- Raspbian Trixie support
+
+### Changed
+- updated auto-installer.sh script to version 1.2
+- config template updated
+
 ## [1.3.1] - 2025-11-25
 
 ### Changed